refactor(auto): encapsulate deduplication to resolve spec private method bypasses
This addresses PR feedback by:
- Creating a SemanticHtml::Deduplicator collaborator to encapsulate sorting, caching, and nested container collapsing.
- Replacing instance_eval/instance_exec calls in spec/lib/html2rss/auto_source/scraper/semantic_html_ranking_spec.rb with direct assertions against the public Deduplicator interface.
- Fixing a typo ('taging' -> 'tagging') in AGENTS.md.
